move BaseController to base-tool module
