+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/dagger/UiControllerModule.kt
@@ -5,8 +5,8 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import dagger.multibindings.IntoMap
+import org.cru.godtools.base.tool.ui.controller.BaseController
 import org.cru.godtools.tract.ui.controller.AnimationController
-import org.cru.godtools.tract.ui.controller.BaseController
 import org.cru.godtools.tract.ui.controller.ButtonController
 import org.cru.godtools.tract.ui.controller.FallbackController
 import org.cru.godtools.tract.ui.controller.FormController
