+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/ui/controller/FallbackController.kt
@@ -5,6 +5,7 @@ import android.view.ViewGroup
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
+import org.cru.godtools.base.tool.ui.controller.BaseController
 import org.cru.godtools.tract.databinding.TractContentFallbackBinding
 import org.cru.godtools.xml.model.Fallback
