+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/ui/controller/InputController.kt
@@ -8,6 +8,7 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import org.cru.godtools.base.model.Event
+import org.cru.godtools.base.tool.ui.controller.BaseController
 import org.cru.godtools.tract.databinding.TractContentInputBinding
 import org.cru.godtools.xml.model.Input
