+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/ui/controller/RestructureUtils.kt
@@ -1,5 +1,6 @@
 package org.cru.godtools.tract.ui.controller
 
+import org.cru.godtools.base.tool.ui.controller.BaseController
 import org.cru.godtools.xml.model.Base
 
 // TODO: move this back into BaseController and make modelClass private again
