+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/ui/controller/TabsController.kt
@@ -11,6 +11,7 @@ import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import org.ccci.gto.android.common.material.tabs.setBackgroundTint
 import org.cru.godtools.base.model.Event
+import org.cru.godtools.base.tool.ui.controller.BaseController
 import org.cru.godtools.tract.databinding.TractContentTabsBinding
 import org.cru.godtools.xml.model.Tabs
 import org.cru.godtools.xml.model.primaryColor
