+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/ui/controller/tips/InlineTipController.kt
@@ -5,8 +5,8 @@ import android.view.ViewGroup
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
+import org.cru.godtools.base.tool.ui.controller.BaseController
 import org.cru.godtools.tract.databinding.TractContentInlineTipBinding
-import org.cru.godtools.tract.ui.controller.BaseController
 import org.cru.godtools.xml.model.tips.InlineTip
 
 internal class InlineTipController private constructor(
